### 2023-04-07 #5349 @editorialbot
[PRE REVIEW]: PAS: a Python Anesthesia Simulator for drug control
**Submitting author:** @BobAubouin (<a href="http://orcid.org/0000-0003-0029-438X">Bob Aubouin--Pairault</a>)
**Repository:** https://github.com/BobAubouin/Python_Anesthesia_Simulator
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.4
