+++ README.md
@@ -1,3 +1,5 @@
+[![status](https://joss.theoj.org/papers/61d34ad9ef855a128509b4279e2c9325/status.svg)](https://joss.theoj.org/papers/61d34ad9ef855a128509b4279e2c9325)
+
 # Python_Anesthesia_Simulator
 The Python Anesthesia Simulator (PAS) models the effect of drugs on physiological variables during total intravenous anesthesia. It is particularly dedicated to the control community, to be used as a benchmark for the design of multidrug controllers. The available drugs are Propofol, Remifentanil, and Norepinephrine, the outputs are the Bispectral index (BIS), Mean Arterial Pressure (MAP), Cardiac Output (CO), and Tolerance of Laryngoscopy (TOL). PAS includes different well-known models along with their uncertainties to simulate inter-patient variability. Blood loss can also be simulated to assess the controller's performance in a shock scenario. Finally, PAS includes standard disturbance profiles and metrics computation to facilitate the evaluation of the controller's performances.
 
@@ -38,7 +40,7 @@ Available in the _Documentation.pdf_ file.
 _GNU General Public License 3.0_
 
 ## Project status
-Ready for JOSS submition
+Submitted to JOSS
 
 ## Author
 Bob Aubouin--Paitault
