### 2023-04-17 #5349 @Kevin-Mattheus-Moerman
@BobAubouin thanks for this submission. I am the AEiC on this track and here to help process the initial steps.

Note, [this PR](https://github.com/BobAubouin/Python_Anesthesia_Simulator/pull/12) should fix the above paper compilation issue.
