### 2023-04-19 b7228ea
fix issue #15 random values of dynamix matrix
