### 2023-04-19 9782754
add casadi and pandas as package dependencies
