### 2023-04-19 34e86ba
fix issue #13 switch coefficient in matrix A
