### 2023-04-22 #5349 @ppxasjsm
Hi @BobAubouin, I am here to help facilitate the editorial process. I am looking for reviewers now, in case you have suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
