### 2023-04-24 30ecf30
Add the fast simu method
