### 2023-04-25 #5349 @ppxasjsm
@BobAubouin could you in the meantime while we are gathering reviewers, address the missing DOI error you get with your paper submission.

It's about this missing DOI: - 10.1016/c2017-0-03401-8 may be a valid DOI for title: Automated Drug Delivery in Anesthesia

Thank you!
