### 2023-04-26 #5349 @BobAubouin
@ppxasjsm I can suggest wahlquisty which works on the subject. Then looking at the list of people who already review for JOSS, I can also suggest shravantata, askuric, JHartzer, gmxavier, ayush9pandey, jbcaillau, sea-bass, vmatos, and Jgoldfar. They are not specific about anesthesia but are on the topic of control.

I will address the missing DOI now!
