+++ paper/bibli.bib
@@ -120,7 +120,8 @@
   title = {Automated Drug Delivery in Anesthesia},
   author = {Copot, Dana},
   year = {2020},
-  publisher = {{Academic Press}}
+  publisher = {{Academic Press}},
+  doi = {10.1016/c2017-0-03401-8}
 }
 
 @article{dewitEffectPropofolHaemodynamics2016,
