+++ src/python_anesthesia_simulator/__init__.py
@@ -1,3 +1,5 @@
 from .simulator import Patient
 from .disturbances import compute_disturbances
 from .metrics import compute_control_metrics
+from .pk_models import CompartmentModel
+from .pd_models import BIS_model, TOL_model, Hemo_PD_model
