+++ paper/bibli.bib
@@ -474,8 +474,8 @@
   file = {/home/aubouinb/Zotero/storage/924XN9Z8/Monnet et al_2011_Norepinephrine increases cardiac preload and reduces preload dependency.pdf;/home/aubouinb/Zotero/storage/95WBD6HJ/Norepinephrine_increases_cardiac_preload_and.12.html}
 }
 
-@book{pardoBasicsAnesthesiaEBook2017,
-  title = {Basics of {{Anesthesia E-Book}}},
+@book{pardoBasicsAnesthesia2017,
+  title = {Basics of {{Anesthesia}}},
   author = {Pardo, Manuel and Miller, Ronald D.},
   year = {2017},
   month = jun,
