### 2023-05-12 badfa74
add noise filter for BIS
