### 2023-05-20 #5480 @editorialbot
[REVIEW]: PAS: a Python Anesthesia Simulator for drug control
**Submitting author:** @BobAubouin (<a href="http://orcid.org/0000-0003-0029-438X">Bob Aubouin--Pairault</a>)
**Repository:** https://github.com/BobAubouin/Python_Anesthesia_Simulator
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
