+++ README.md
@@ -21,7 +21,7 @@ Documentation is available in the _Documentation.pdf_ file and different noteboo
 - To report a bug or request a feature please open an issue.
 - To contribute, you can fork the repository and ask for a pull request.
 
-If you want to contact me for any reason, I'm available by mail [mail](bob.aubouin-pairault@gipsa-lab.fr).
+If you want to contact me for any reason, I'm available by [mail](bob.aubouin-pairault@gipsa-lab.fr).
 
 ## Structure
 
