### 2023-06-18 BobAubouin/Python_Anesthesia_Simulator#17 @JHartzer
Improve Code Testing and Coverage
Consider adding a test framework such as [pytest](https://pypi.org/project/pytest/) or [unittest](https://docs.python.org/3/library/unittest.html) and coupling that with a tool such as [coverage](https://pypi.org/project/coverage/). This would help improve understanding of which lines of code are being tested and identify any gaps in testing coverage.
