### 2023-06-18 BobAubouin/Python_Anesthesia_Simulator#18 @JHartzer
Incorporate Feature Work into Issue Tracker
Related to the work in [JOSS Review 5480](https://github.com/openjournals/joss-reviews/issues/5480), I would recommend taking the future work outlined in the paper and incorporating it into this feature tracker to improve visibility for the rest of the community and help indicate where contributions are needed.
