### 2023-06-18 #5480 @JHartzer
In addition to the issue above, I also added [Issue 17](https://github.com/BobAubouin/Python_Anesthesia_Simulator/issues/17) regarding additional testing and coverage. This could improve the robustness of testing and help indicate the amount of testing coverage.

Lastly, I would suggest linking or mentioning the new ECC 2023 publication in the repository once it is available.

None of these are substantial issues, and I imagine the code quality will only improve moving forward. Therefore, I  recommend JOSS acceptance
