### 2023-06-19 #5480 @BobAubouin
Hi @JHartzer, thanks for your review and your suggestions!

I will add the future work into your issue and the citation to the ECC23 into the paper.

About the test framework, I'm not familiar with those, but I will consider it for improvement.
