### 2023-06-19 399caa0
Add ECC23 citation to the paper
