+++ paper/bibli.bib
@@ -15,6 +15,18 @@
   file = {/home/aubouinb/Zotero/storage/XCMUEYMW/The_Determinants_of_Propofol_Induction_of.20.html}
 }
 
+@inproceedings{aubouin--pairaultDatabasedPharmacodynamicModeling2023,
+  title = {Data-Based {{Pharmacodynamic Modeling}} for {{BIS}} and {{Mean Arterial Pressure Prediction}} during {{General Anesthesia}}},
+  booktitle = {{{ECC}} 2023 - 21st {{European Control Conference}}},
+  author = {{Aubouin--Pairault}, Bob and Fiacchini, Mirko and Dang, Thao},
+  year = {2023},
+  month = jun,
+  urldate = {2023-05-22},
+  abstract = {In this paper, a data-based approach is used to predict the effect of Propofol and Remifentanil on Bispectral Index (BIS) and Mean Arterial Pressure (MAP) during total intravenous anesthesia. In particular, we aim to reproduce the measured data by identifying the pharmacodynamic function using machine-learning techniques. Features from the output of classic pharmacokinetic models and patient information are considered. Five learning methods are tested including linear models, support vector machine, Kernel, k-neighbors regressors, and neural-network. Learning and testing are performed on a particular subset of 150 surgery cases extracted from the VitalDB database. Results show that this approach improves the classic surface-response methods for BIS and MAP prediction and can be used for anesthesia control applications.},
+  langid = {english},
+  file = {/home/aubouinb/Zotero/storage/7ZL7IELR/Aubouin--Pairault et al_2023_Data-based Pharmacodynamic Modeling for BIS and Mean Arterial Pressure.pdf}
+}
+
 @article{beloeilNorepinephrineKineticsDynamics2005,
   title = {Norepinephrine Kinetics and Dynamics in Septic Shock and Trauma Patients},
   author = {Beloeil, H. and Mazoit, J.-X. and Benhamou, D. and Duranteau, J.},
