### 2023-06-30 #5480 @BobAubouin
Hi Erhan,

You can access the review on github [here](https://github.com/openjournals/joss-reviews/issues/5480) and edit your comment containing the checklist by clicking on the three dots at the top right of your comment.
