+++ src/python_anesthesia_simulator/pd_models.py
@@ -21,7 +21,7 @@ class BIS_model:
     Parameters
     ----------
     hill_model : str, optional
-        Only 'Bouillon' [9] is available. Ignored if hill_param is specified.
+        Only 'Bouillon' [Bouillon2004]_ is available. Ignored if hill_param is specified.
     hill_param : list, optional
         Parameter of the Hill model (Propo Remi interaction)
         list [c50p_BIS, c50r_BIS, gamma_BIS, beta_BIS, E0_BIS, Emax_BIS]:
@@ -59,10 +59,11 @@ class BIS_model:
 
     References
     ----------
-    .. [9]  T. W. Bouillon et al., “Pharmacodynamic Interaction between Propofol and Remifentanil
+    .. [Bouillon2004]  T. W. Bouillon et al., “Pharmacodynamic Interaction between Propofol and Remifentanil
             Regarding Hypnosis, Tolerance of Laryngoscopy, Bispectral Index, and Electroencephalographic
             Approximate Entropy,” Anesthesiology, vol. 100, no. 6, pp. 1353–1372, Jun. 2004,
             doi: 10.1097/00000542-200406000-00006.
+
     """
 
     def __init__(self, hill_model: str = 'Bouillon', hill_param: list = None,
@@ -150,7 +151,7 @@ class BIS_model:
     def update_param_blood_loss(self, v_ratio: float):
         """Update PK coefficient to mimic a blood loss.
 
-        Update the c50p parameters thanks to the blood volume ratio.
+        Update the c50p parameters thanks to the blood volume ratio. The values are estimated from [Johnson2003]_.
 
         Parameters
         ----------
@@ -161,15 +162,17 @@ class BIS_model:
         -------
         None.
 
-        """
-        # value estimated from K. B. Johnson et al., “The Influence of Hemorrhagic Shock on Propofol: A Pharmacokinetic
-        # and Pharmacodynamic Analysis,” Anesthesiology, vol. 99, no. 2, pp. 409–420, Aug. 2003,
-        # doi: 10.1097/00000542-200308000-00023.
+        References
+        ----------
+        .. [Johnson2003]  K. B. Johnson et al., “The Influence of Hemorrhagic Shock on Propofol: A Pharmacokinetic
+            and Pharmacodynamic Analysis,” Anesthesiology, vol. 99, no. 2, pp. 409–420, Aug. 2003,
+            doi: 10.1097/00000542-200308000-00023.
 
+        """
         self.c50p = self.c50p_init - 3/0.5*(1-v_ratio)
 
     def inverse_hill(self, BIS: float, c_es_remi: float = 0) -> float:
-        """Compute Propofol effect site concentration from BIS and Remifentanil Effect site concentration.
+        """Compute Propofol effect site concentration from BIS and Remifentanil effect site concentration.
 
         Parameters
         ----------
@@ -232,7 +235,7 @@ class TOL_model():
     Parameters
     ----------
     model : str, optional
-        Only 'Bouillon'[9] is available. Ignored if model_param is specified. The default is 'Bouillon'.
+        Only 'Bouillon'[Bouillon2004]_ is available. Ignored if model_param is specified. The default is 'Bouillon'.
     model_param : list, optional
         Model parameters, model_param = [c50p, c50p, gammaP, gammaR, Preopioid intensity].
         The default is None.
@@ -252,12 +255,6 @@ class TOL_model():
     pre_intensity : float
         Preopioid intensity.
 
[…]

[…]
