+++ docs/source/src.python_anesthesia_simulator.rst
@@ -1,0 +1,41 @@
+
+Simulator module
+================
+
+.. automodule:: src.python_anesthesia_simulator.simulator
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+PK models module
+================
+
+.. automodule:: src.python_anesthesia_simulator.pk_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+PD models module
+================
+
+.. automodule:: src.python_anesthesia_simulator.pd_models
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Disturbances module
+===================
+
+.. automodule:: src.python_anesthesia_simulator.disturbances
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Metrics module
+==============
+
+.. automodule:: src.python_anesthesia_simulator.metrics
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
