+++ .readthedocs.yaml
@@ -20,8 +20,8 @@ sphinx:
    configuration: docs/source/conf.py
 
 # Optionally build your docs in additional formats such as PDF and ePub
-formats:
-	- pdf
+#formats:
+# - pdf
 
 
 # Optional but recommended, declare the Python requirements required
