### 2023-07-03 7c9f41f
re-fix doc requirements
