+++ docs/requirements.txt
@@ -1,2 +1,6 @@
 numpydoc
 sphinx_rtd_theme
+casadi==3.5.5
+pandas==1.4.3
+numpy==1.22.3
+matplotlib==3.5.2
