+++ .readthedocs.yaml
@@ -29,4 +29,4 @@ sphinx:
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
    install:
-   - requirements: docs/requirements.txt, requirements.txt
+   - requirements: docs/requirements.txt
