+++ requirements.txt
@@ -3,5 +3,6 @@ pandas==1.4.3
 bokeh==2.4.3
 numpy==1.22.3
 matplotlib==3.5.2
+control==0.9.3.post2
 pyswarm==0.6
 
