+++ docs/requirements.txt
@@ -4,3 +4,4 @@ casadi==3.5.5
 pandas==1.4.3
 numpy==1.22.3
 matplotlib==3.5.2
+control==0.9.3.post2
