### 2023-07-03 bffbe8f
update docs with examples
