+++ docs/source/examples/pid_sim_functions.py
@@ -1,0 +1,221 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Apr 25 14:36:09 2022
+
+@author: aubouinb
+"""
+# Standard import
+import time
+import multiprocessing
+from functools import partial
+
+# Third party imports
+import numpy as np
+import pandas as pd
+import casadi as cas
+from pyswarm import pso
+
+# Local imports
+import python_anesthesia_simulator as pas
+
+# Patient table:
+# index, Age, H[cm], W[kg], Gender, Ce50p, Ce50r, γ, β, E0, Emax
+Patient_table = [[1,  40, 163, 54, 0, 4.73, 24.97,  2.97,  0.3, 97.86, 89.62],
+                 [2,  36, 163, 50, 0, 4.43, 19.33,  2.04,  0.29, 89.1, 98.86],
+                 [3,  28, 164, 52, 0, 4.81, 16.89,  1.18,  0.14, 93.66, 94.],
+                 [4,  50, 163, 83, 0, 3.86, 20.97,  1.08,  0.12, 94.6, 93.2],
+                 [5,  28, 164, 60, 1, 5.22, 18.95,  2.43,  0.68, 97.43, 96.21],
+                 [6,  43, 163, 59, 0, 3.41, 23.26,  3.16,  0.58, 85.33, 97.07],
+                 [7,  37, 187, 75, 1, 4.83, 15.21,  2.94,  0.13, 91.87, 90.84],
+                 [8,  38, 174, 80, 0, 4.36, 13.86,  3.37,  1.05, 97.45, 96.36],
+                 [9,  41, 170, 70, 0, 4.57, 16.2,  1.55,  0.16, 85.83, 94.6],
+                 [10, 37, 167, 58, 0, 6.02, 23.47,  2.83,  0.77, 95.18, 88.17],
+                 [11, 42, 179, 78, 1, 3.79, 22.25,  2.35,  1.12, 98.02, 96.95],
+                 [12, 34, 172, 58, 0, 5.7, 18.64,  2.67,  0.4, 99.57, 96.94],
+                 [13, 38, 169, 65, 0, 4.64, 19.50,  2.38,  0.48, 93.82, 94.40]]
+
+
+def simu(Patient_info: list, style: str, PID_param: list,
+         random_PK: bool = False, random_PD: bool = False) -> tuple[float, list, list]:
+    """
+    Perform a closed-loop Propofol-Remifentanil anesthesia simulation with a PID controller.
+
+    Parameters
+    ----------
+    Patient_info : list
+        list of patient informations, Patient_info = [Age, H[cm], W[kg], Gender, Ce50p, Ce50r, γ, β, E0, Emax].
+    style : str
+        Either 'induction' or 'total' to describe the phase to simulate.
+    PID_param : list
+        Parameters of the NMPC controller, MPC_param = [Kp, Ti, Td, ratio].
+    random_PK : bool, optional
+        Add uncertaintie to the PK model. The default is False.
+    random_PD : bool, optional
+        Add uncertainties to the PD model. The default is False.
+
+    Returns
+    -------
+    IAE : float
+        Integrated Absolute Error, performance index of the function
+    data : list
+        list of the signals during the simulation, data = [BIS, MAP, CO, up, ur]
+    BIS_param: list
+        BIS parameters of the simulated patient.
+
+    """
+
+    hill_param = Patient_info[4:]
+    if hill_param[0] is None:
+        hill_param = None
+    George = pas.Patient(Patient_info[:4], hill_param=hill_param,
+                         random_PK=random_PK, random_PD=random_PD)
[…]
