+++ docs/requirements.txt
@@ -1,7 +1,10 @@
 numpydoc
 sphinx_rtd_theme
+nbsphinx
 casadi==3.5.5
 pandas==1.4.3
 numpy==1.22.3
 matplotlib==3.5.2
 control==0.9.3.post2
+pyswarm==0.6
+bokeh
