+++ paper/bibli.bib
@@ -8,7 +8,6 @@
   number = {3},
   pages = {656},
   issn = {0003-2999},
-  doi = {10.1213/00000539-200103000-00020},
   urldate = {2023-02-14},
   abstract = {Recently it was reported that the pharmacokinetics of propofol are modified by changes in cardiac output. The objective of this study was to evaluate the effects of cardiac output and other factors on the hypnotic dose of propofol. One-hundred surgical patients were administered indocyanine green immediately before the induction of anesthesia to measure their cardiac outputs and blood volumes. Propofol (250 {$\mu$}g {$\cdot$} kg-1 {$\cdot$} min-1) was infused IV for 8 min, and the hypnotic dose of propofol and the time to hypnosis were recorded. The plasma concentration of propofol immediately after 2 mg/kg infusion was measured. Multiple regression analysis showed that, in addition to age and weight, cardiac output was a small but significant factor for predicting the hypnotic dose of propofol (R2 = 0.468, P {$<$} 0.001), the time to hypnosis (R2 = 0.454, P {$<$} 0.001), and the plasma concentration of propofol (R2 = 0.248, P {$<$} 0.01). Cardiac output, age, and weight showed similar partial coefficients for the hypnotic dose (0.128, 0.137, and 0.140, respectively).},
   langid = {american},
@@ -37,7 +36,6 @@
   number = {6},
   pages = {782--788},
   issn = {0007-0912},
-  doi = {10.1093/bja/aei259},
   urldate = {2023-02-22},
   abstract = {Background. There is considerable variability in the inter-patient response to norepinephrine. Pharmacokinetic studies of dopamine infusion in volunteers and in patients have also shown large variability. The purpose of this study was to define the pharmacokinetics of norepinephrine in septic shock and trauma patients.Methods. After Ethical Committee approval and written informed family consent, 12 patients with septic shock and 11 trauma patients requiring norepinephrine infusion were studied. Norepinephrine dose was increased in three successive steps of 0.1 mg kg-1 min-1 at 15-min intervals (20\% maximum allowed increase in arterial pressure). Arterial blood was sampled before and at 0.5, 13, and 15 min after each infusion rate change and 30 s, 1, 2, 5, 10, and 15 min after return to baseline dosing. Norepinephrine was assayed by HPLC. The pharmacokinetics were modelled using NONMEM (one-compartment model). The effects of group, body weight (BW), gender and SAPS II (Simplified Acute Physiology Score II) [Le Gall JR, Lemeshow S, Saulnier F. A new Simplified Acute Physiology Score (SAPS II) based on a European/North American multicenter study. J Am Med Assoc 1993; 270: 2957\textendash 63] patients score on clearance (CL) and volume of distribution (V) were tested.Results. Group, gender, and BW did not influence CL or V. CL was negatively related to SAPS II. CL and T1/2 varied from 3 litre min-1 and 2 min, respectively, when SAPS II=20 to 0.9 litre min-1 and 6.8 min when SAPS II=60.Conclusion. In trauma patients and in septic shock patients, norepinephrine clearance is negatively related to SAPS II.},
   file = {/home/aubouinb/Zotero/storage/7FP5MSCM/Beloeil et al_2005_Norepinephrine kinetics and dynamics in septic shock and trauma patients.pdf;/home/aubouinb/Zotero/storage/7LJV628N/257220.html}
@@ -53,7 +51,6 @@
   number = {6},
   pages = {583--596},
   issn = {1573-8744},
-  doi = {10.1007/s10928-020-09712-1},
   urldate = {2022-01-17},
[…]
