+++ docs/requirements.txt
@@ -1,10 +1,10 @@
 numpydoc
 sphinx_rtd_theme
 nbsphinx
-casadi==3.5.5
+casadi==3.6.3
 pandas==1.4.3
-numpy==1.22.3
-matplotlib==3.5.2
-control==0.9.3.post2
+numpy==1.25.0
+matplotlib==3.7.1
+control==0.9.4
 pyswarm==0.6
-bokeh
+bokeh==3.2.0
