### 2023-07-04 #5480 @BobAubouin
Hi Erhan,

Thanks for your feedback! I have been working on the documentation, it is now available online here: [https://python-anesthesia-simulator.readthedocs.io](https://python-anesthesia-simulator.readthedocs.io). It includes the statement of need as introduction and also the examples.
