+++ docs/source/index.rst
@@ -19,7 +19,7 @@ understand how to use it.
 
 
 .. toctree::
-   :maxdepth: 5
+   :maxdepth: 2
    :caption: Contents:
 
    intro
