+++ docs/source/examples.rst
@@ -8,8 +8,6 @@ The source code for the examples below are available in the `docs/source/example
 subdirecory of the source code distribution.  The codes can also be accessed online
 via the `PAS GitHub repository <https://github.com/BobAubouin/Python_Anesthesia_Simulator/tree/main/docs/source/examples>`_.
 
-Jupyter notebooks
-=================
 
 The examples below use `PAS` in a Jupyter notebook environment.
 These notebooks demonstrate the use of the toolbox for different applications.
