### 2023-07-05 BobAubouin/Python_Anesthesia_Simulator#18 @BobAubouin
It is done in Issue #19
