### 2023-07-05 #5480 @BobAubouin
@ppxasjsm  I'm good with the repo and the manuscript! I added a tag v1.0.0 to the current git version since it have changed a bit since the submission.
