### 2023-07-08 #5480 @ppxasjsm
@BobAubouin I have now gone through your paper version, can you please make the following small fixes:

- [ ] first paragraph performances -> performance
- [ ] drug names: are they branded? Then capitalised otherwise if they are generic do not capitalise
- [ ] Bispectral index, mean arterial pressure, cardiac output,  and tol should not be capitalised, the acronyms should be in capital.
- [ ] The statement of need is first discussed -> replace the comma with a semicolon

Statement of need:
- [ ] which has drawn -> that has attracted
- [ ] an arduous work: either an arduous task or arduous work
- [ ] propositions -> suggestions
- [ ] If this is a first step -> if this is the first step
- [ ] are still not -> are not yet
- [ ] performances -> performance
- [ ] the control community has been historically using  -> has historically used
- [ ] facilitate -> facilitates
- [ ] simulator at the maintenance phase -> in the maintenance phase
Pharmacokinetics:
- [ ] compartments -> compartment
- [ ] 3-compartments -> 3-compartment
- [ ] drugs -> drugs'
- [ ] made for the Propofol -> made for propofol
- [ ] impact the PK system -> affect the PK
Pharmacodynamics:
- [ ] Then a Hill curve is -> A Hill curve was then used...
- [ ] For Propofol the value from [...] are-> For propofol values [...] are
- [ ] Thus -> Thus,
- [ ] the crude assumption [...] has been made ->  was made
- [ ] behaviour -> behavior
- [ ] assumption however -> assumption, however
Conclusion and future work:
- [ ]  to design and test multidrug -> for designing and testing multidrug controllers in [...]
- [ ] The Respiratory system -> respiratory
- [ ] MAP effect, and, experimental -> MAP effect, and the experimental

I am not sure what happened with the DOIs, they all appear to be missing but had not been missing previously can you make sure all the DOIs are available please?
