### 2023-07-08 #5480 @ppxasjsm
@BobAubouin can you please also check the following:
- [ ] Reference Fairfield1991 appears to be in all caps letters
