### 2023-07-12 #5480 @ppxasjsm
@BobAubouin, I am not sure what happened with the DOIs, but they all seem to be missing, but weren't in the past. Can you please make sure all the DOIs that can be added to the references will be and none are missing. You can check these with editorialbot check references after you generated a new version of the pdf with editorialbot generate pdf.
