### 2023-07-20 b1e4812
small fix in paper
