+++ paper/PAS_paper.md
@@ -31,26 +31,26 @@ bibliography: bibli.bib
 
 # Summary
 
-The Python Anesthesia Simulator (PAS) models the effect of drugs on physiological variables during total intravenous anesthesia. It is particularly dedicated to the control community, to be used as a benchmark for the design of multidrug controllers. The available drugs are Propofol, Remifentanil, and Norepinephrine, the outputs are the Bispectral index (BIS), Mean Arterial Pressure (MAP), Cardiac Output (CO), and Tolerance of Laryngoscopy (TOL). PAS includes different well-known models along with their uncertainties to simulate inter-patient variability. Blood loss can also be simulated to assess the controller's performance in a shock scenario. Finally, PAS includes standard disturbance profiles and metric computation to facilitate the evaluation of the controller's performances. The statement of need of this package is first discussed, then some pieces of information about the programmed models are provided to the reader. The future developments of the package are discussed at the end.
+The Python Anesthesia Simulator (PAS) models the effect of drugs on physiological variables during total intravenous anesthesia. It is particularly dedicated to the control community, to be used as a benchmark for the design of multidrug controllers. The available drugs are propofol, remifentanil, and norepinephrine, the outputs are the bispectral index (BIS), mean arterial pressure (MAP), cardiac output (CO), and tolerance of laryngoscopy (TOL). PAS includes different well-known models along with their uncertainties to simulate inter-patient variability. Blood loss can also be simulated to assess the controller's performance in a shock scenario. Finally, PAS includes standard disturbance profiles and metric computation to facilitate the evaluation of the controller's performance. The statement of need of this package is first discussed; then some pieces of information about the programmed models are provided to the reader. The future developments of the package are discussed at the end.
 
 # Statement of need
 
-Closing the loop for drug dosage during general anesthesia is a challenging task which has drawn the attention of the control community for more than two decades. In fact, the high need for reliability coupled with a highly uncertain system makes the design of a controller an arduous work. Numerous closed-loop control strategies have been proposed and reviewed in @ilyasReviewModernControl2017, @copotAutomatedDrugDelivery2020, @singhArtificialIntelligenceAnesthesia2022, and @ghitaClosedLoopControlAnesthesia2020 for instance. 
+Closing the loop for drug dosage during general anesthesia is a challenging task that has attracted the attention of the control community for more than two decades. In fact, the high need for reliability coupled with a highly uncertain system makes the design of a controller arduous work. Numerous closed-loop control strategies have been proposed and reviewed in @ilyasReviewModernControl2017, @copotAutomatedDrugDelivery2020, @singhArtificialIntelligenceAnesthesia2022, and @ghitaClosedLoopControlAnesthesia2020 for instance. 
 
 
-Since it is a long process to clinically test a control method for drug injection, many papers only rely on simulations to demonstrate the advantages of their propositions. However, two meta-studies @brogiClinicalPerformanceSafety2017, @puriMulticenterEvaluationClosedLoop2016 agree on the fact that automated control during anesthesia brings better stability to the physiological signals during the procedure. Since the methods that were first tested on simulation led to good performances in clinical situations, it could be concluded that simulations are realistic enough to be used as a first test.
[…]
