+++ paper/bibli.bib
@@ -662,7 +662,7 @@
   number = {1},
   pages = {86--93},
   issn = {1658-354X},
-  doi = {10.4103/sja.sja\_669\_21},
+  doi = {10.4103/sja.sja_669_21},
   urldate = {2022-09-22},
   abstract = {Rapid advances in Artificial Intelligence (AI) have led to diagnostic, therapeutic, and intervention-based applications in the field of medicine. Today, there is a deep chasm between AI-based research articles and their translation to clinical anesthesia, which needs to be addressed. Machine learning (ML), the most widely applied arm of AI in medicine, confers the ability to analyze large volumes of data, find associations, and predict outcomes with ongoing learning by the computer. It involves algorithm creation, testing and analyses with the ability to perform cognitive functions including association between variables, pattern recognition, and prediction of outcomes. AI-supported closed loops have been designed for pharmacological maintenance of anesthesia and hemodynamic management. Mechanical robots can perform dexterity and skill-based tasks such as intubation and regional blocks with precision, whereas clinical-decision support systems in crisis situations may augment the role of the clinician. The possibilities are boundless, yet widespread adoption of AI is still far from the ground reality. Patient-related ``Big Data'' collection, validation, transfer, and testing are under ethical scrutiny. For this narrative review, we conducted a PubMed search in 2020-21 and retrieved articles related to AI and anesthesia. After careful consideration of the content, we prepared the review to highlight the growing importance of AI in anesthesia. Awareness and understanding of the basics of AI are the first steps to be undertaken by clinicians. In this narrative review, we have discussed salient features of ongoing AI research related to anesthesia and perioperative care.},
   pmcid = {PMC8846233},
