### 2023-07-20 #5480 @BobAubouin
Hi @ppxasjsm thanks for the feedbacks.

I have completed all the small fixes and the DOI problem seems to be solved!
