### 2023-07-21 #5480 @BobAubouin
The release have been archive on Zenodo with the following DOI: 10.5281/zenodo.8171326.
It is available [here](https://zenodo.org/record/8171326).
