+++ paper/PAS_paper.md
@@ -41,7 +41,7 @@ Closing the loop for drug dosage during general anesthesia is a challenging task
 Since it is a long process to clinically test a control method for drug injection, many papers only rely on simulations to demonstrate the advantages of their suggestions. However, two meta-studies @brogiClinicalPerformanceSafety2017, @puriMulticenterEvaluationClosedLoop2016 agree on the fact that automated control during anesthesia brings better stability to the physiological signals during the procedure. Since the methods that were first tested on simulation led to good performances in clinical situations, it could be concluded that simulations are realistic enough to be used as a first test.
 
 
-In addition to being significantly faster than clinical tests, simulations also allow researchers to use the same framework to compare their methods. Up to now, many different controllers have been proposed and tested on their own frameworks and it is often hard to compare them properly. This is aggravated by the fact that the code of the algorithms is rarely open-source which makes hard the reproducibility of the results (to the author's knowledge, @comanAnesthesiaGUIDEMATLABTool2022 is the only open-source controller for anesthesia). Ionescu *et al.* recently address this issue by providing an open-source anesthesia simulator in @ionescuOpenSourcePatient2021a. If this is the first step to a FAIR (Findable, Accessible, Interoperable, Reusable) science, it can be improved. In fact, the measurement noise and disturbance scenarios are not yet specified. Moreover, while the simulator could incorporate them, the level of uncertainty to take into account in the patient model is also not fixed. Finally, the metrics used to compare the performance must be explicit and used by all users to compare the proposed control methods.
+In addition to being significantly faster than clinical tests, simulations also allow researchers to use the same framework to compare their methods. Up to now, many controllers have been proposed and tested on their own frameworks, and it is often hard to compare them properly. This is aggravated by the fact that the code of the algorithms is rarely open-source which makes hard the reproducibility of the results (to the author's knowledge, @comanAnesthesiaGUIDEMATLABTool2022 is the only open-source controller for anesthesia). Ionescu *et al.* recently address this issue by providing an open-source anesthesia simulator in @ionescuOpenSourcePatient2021a. If this is the first step to a FAIR (Findable, Accessible, Interoperable, Reusable) science, it can be improved. In fact, the measurement noise and disturbance scenarios are not yet specified. Moreover, while the simulator could incorporate them, the level of uncertainty to take into account in the patient model is also not fixed. Finally, the metrics used to compare the performance must be explicit and used by all users to compare the proposed control methods.
   
 
 With the Python Anesthesia Simulator, a full pipeline is proposed to test control methods for drug dosage in anesthesia. By using Python, an open-source language, we hope that everyone will be able to use our simulator. If the control community has historically used Matlab, the use of the control package [@fullerPythonControlSystems2021] in PAS facilitates the transition, with Matlab compatibility functions. Moreover, from an interdisciplinary point of view, Python is already widely used by the data science and machine learning community. On the anesthesia topic, the database VitalDB [@leeVitalRecorderFree2018] has a Python API and could be used together with PAS to develop data-based approaches, as already done in @aubouinpairaultDatabasedPharmacodynamicModeling2023.
@@ -66,16 +66,16 @@ In PAS all drug effects are described by the well-know Pharmacokinetic-Pharmacod
 
 ## Pharmacokinetics
 
[…]
