### 2023-08-06 #5480 @ppxasjsm
@BobAubouin, I am sorry this ended up taking much longer than I thought. We are ready to accept, there is just one DOI that should be added to the paper. This is for the title: Data-Based Pharmacodynamic Modeling for BIS and Mean Arterial Pressure Prediction during General Anesthesia

Looking at the link: https://ieeexplore.ieee.org/document/10178214, the DOI suggest and just stated below should be added:
10.23919/ECC57647.2023.10178214

Thanks a lot! As soon as this is done I can recommend-accept.
