+++ paper/bibli.bib
@@ -21,6 +21,7 @@
   author = {Aubouin Pairault, Bob and Fiacchini, Mirko and Dang, Thao},
   year = {2023},
   month = jun,
+  doi = {10.23919/ecc57647.2023.10178214},
   urldate = {2023-05-22},
   abstract = {In this paper, a data-based approach is used to predict the effect of Propofol and Remifentanil on Bispectral Index (BIS) and Mean Arterial Pressure (MAP) during total intravenous anesthesia. In particular, we aim to reproduce the measured data by identifying the pharmacodynamic function using machine-learning techniques. Features from the output of classic pharmacokinetic models and patient information are considered. Five learning methods are tested including linear models, support vector machine, Kernel, k-neighbors regressors, and neural-network. Learning and testing are performed on a particular subset of 150 surgery cases extracted from the VitalDB database. Results show that this approach improves the classic surface-response methods for BIS and MAP prediction and can be used for anesthesia control applications.},
   langid = {english},
