### 2023-08-10 #5480 @Kevin-Mattheus-Moerman
@BobAubouin I have checked this review, the archive, the repository, and the paper and most seems in order. We only have the below minor points that need your attention:

- [ ] Please edit your archive license (currently `GNU Affero General Public License v3.0`) to match your software license (`GNU General Public License v3.0`)
- [ ] The name `richmond` should likely feature a capital letter so be written as `Richmond`
- [ ] Recommendation: In your affiliations you may spell out `Université Grenoble Alpes` (no need to use the shortened `Univ`). Also if `38000` refers to a postal code, you may choose to leave that out.
