### 2023-08-14 #5480 @BobAubouin
Hi @Kevin-Mattheus-Moerman, thanks for this final check!

 I just changed the archive license to "GNU General Public License v3.0 only", I'm not sure if this ok but "GNU General Public License v3.0" was not in the list of available licenses.

I also add a capital letter to "Richmond".

About the affiliation, this is the exact sentence imposed by my lab, I prefer to keep it as it is.
